+++ python/GafferUI/__init__.py
@@ -269,6 +269,7 @@ from . import Examples
 from .NameValuePlugValueWidget import NameValuePlugValueWidget
 from .ShufflePlugValueWidget import ShufflePlugValueWidget
 from .ShufflePlugValueWidget import ShufflesPlugValueWidget
+from .BackgroundTaskDialogue import BackgroundTaskDialogue
 
 # and then specific node uis