+++ src/Gaffer/ScriptNode.cpp
@@ -212,9 +212,12 @@ std::string readFile( const std::string &fileName )
 		throw IECore::IOException( "Unable to open file \"" + fileName + "\"" );
 	}
 
+	const IECore::Canceller *canceller = Context::current()->canceller();
+
 	std::string s;
 	while( !f.eof() )
 	{
+		IECore::Canceller::check( canceller );
 		if( !f.good() )
 		{
 			throw IECore::IOException( "Failed to read from \"" + fileName + "\"" );